Review feedback,
- Move some tests around and revert Exposed fns to be internal
- Cleanup the ToReconcileable fn interface, doesn't error anymore, we
  don't use the list value it returns
- More explicit error handling for looking up inventory
