+++ core/server/sync_test.go
@@ -88,7 +88,7 @@ func TestSync(t *testing.T) {
 			WithSource: true,
 		},
 		reconcilable: fluxsync.HelmReleaseAdapter{HelmRelease: hr},
-		source:       fluxsync.NewReconcileable(helmRepo),
+		source:       fluxsync.HelmRepositoryAdapter{HelmRepository: helmRepo},
 	}, {
 		name: "kustomization no source",
 		msg: &pb.SyncFluxObjectRequest{
@@ -105,7 +105,7 @@ func TestSync(t *testing.T) {
 			WithSource: true,
 		},
 		reconcilable: fluxsync.KustomizationAdapter{Kustomization: kust},
-		source:       fluxsync.NewReconcileable(gitRepo),
+		source:       fluxsync.GitRepositoryAdapter{GitRepository: gitRepo},
 	}, {
 		name: "gitrepository",
 		msg: &pb.SyncFluxObjectRequest{
@@ -171,7 +171,7 @@ func TestSync(t *testing.T) {
 			WithSource: true,
 		},
 		reconcilable: fluxsync.HelmReleaseAdapter{HelmRelease: hr},
-		source:       fluxsync.NewReconcileable(helmRepo),
+		source:       fluxsync.HelmRepositoryAdapter{HelmRepository: helmRepo},
 	}}
 
 	for _, tt := range tests {
